+++ include/ElementReproducingKernel.h
@@ -7,9 +7,9 @@
 #include <string>
 #include <vector>
 
+#include "ComputeForceSmoothedCell.h"
 #include "Constants.h"
 #include "ElementBase.h"
-#include "ElementForceProcessor.h"
 #include "FieldData.h"
 #include "FunctionValueStorageProcessor.h"
 #include "Kokkos_Core.hpp"
@@ -60,7 +60,7 @@ class ElementReproducingKernel : public ElementBase {
         if (m_material) {
             has_state = m_material->HasState();
         }
-        m_element_processor = std::make_shared<ElementForceProcessor<1>>(m_mesh_data, m_field_query_data_gather[0].name, force_field_name, m_part_names, has_state);
+        m_compute_force = std::make_shared<aperi::ComputeForceSmoothedCell>(m_mesh_data, m_field_query_data_gather[0].name, force_field_name, m_part_names, has_state);
     }
 
     void FindNeighbors() {
@@ -107,11 +107,11 @@ class ElementReproducingKernel : public ElementBase {
      */
     void ComputeInternalForceAllElements(double time_increment) override {
         assert(this->m_material != nullptr);
-        assert(m_element_processor != nullptr);
+        assert(m_compute_force != nullptr);
 
         // Loop over all elements and compute the internal force
-        m_element_processor->UpdateFields();  // Updates the ngp fields
-        m_element_processor->for_each_cell_gather_scatter_nodal_data<Material::StressFunctor>(*m_smoothed_cell_data, this->m_material->GetStressFunctor());
+        m_compute_force->UpdateFields();  // Updates the ngp fields
+        m_compute_force->ForEachCellComputeForce(*m_smoothed_cell_data, this->m_material->GetStressFunctor());
     }
 
     /**
@@ -128,7 +128,7 @@ class ElementReproducingKernel : public ElementBase {
     const std::vector<std::string> m_part_names;
     std::shared_ptr<aperi::MeshData> m_mesh_data;
     double m_kernel_radius_scale_factor;
-    std::shared_ptr<aperi::ElementForceProcessor<1>> m_element_processor;
+    std::shared_ptr<aperi::ComputeForceSmoothedCell> m_compute_force;
     std::shared_ptr<aperi::SmoothedCellData> m_smoothed_cell_data;
     bool m_use_one_pass_method;
 };
